refactor: uplift user avatar and profile link components
