+++ projects/client/src/lib/sections/lists/components/UserAvatar.svelte
