+++ projects/client/src/lib/sections/lists/social/SocialActivityItem.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import type { SocialActivity } from "$lib/requests/models/SocialActivity";
-  import UserAvatar from "$lib/sections/summary/components/_internal/UserAvatar.svelte";
-  import UserProfileLink from "$lib/sections/summary/components/_internal/UserProfileLink.svelte";
+  import UserAvatar from "../components/UserAvatar.svelte";
+  import UserProfileLink from "../components/UserProfileLink.svelte";
   import SocialActivityCard from "./SocialActivityCard.svelte";
   import SocialActivitySummaryCard from "./SocialActivitySummaryCard.svelte";
